# docs/swagger.yaml
basePath: /
definitions:
  handlers.CreateTodoRes:
    properties:
      todo_id:
        type: string
    type: object
  handlers.UpdateOrDeleteTodoRes:
    properties:
      message:
        type: string
    type: object
  models.CreateTodoDTO:
    properties:
      completed:
        type: boolean
      date:
        type: string
      description:
        maxLength: 50
        minLength: 5
        type: string
      title:
        maxLength: 50
        minLength: 5
        type: string
    required:
    - completed
    - date
    - description
    - title
    type: object
  models.Todo:
    properties:
      completed:
        type: boolean
      date:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  models.UpdateTodoDTO:
    properties:
      completed:
        type: boolean
      date:
        type: string
      description:
        maxLength: 50
        minLength: 5
        type: string
      title:
        maxLength: 50
        minLength: 5
        type: string
    required:
    - completed
    - date
    - description
    - title
    type: object
host: localhost:8080
info:
  contact:
    name: Ben Davis
  description: An example template of a Golang backend API using Echo and MongoDB
  license:
    name: MIT
  title: The Better Backend Template
  version: "0.1"
paths:
  /health:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - text/plain
      responses:
        "200":
          description: OK
      summary: Show the status of server.
      tags:
      - health
  /todos:
    get:
      consumes:
      - '*/*'
      description: fetch every todo available.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Todo'
            type: array
      summary: Get all todos.
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: create a single todo.
      parameters:
      - description: Todo to create
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.CreateTodoDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CreateTodoRes'
      summary: Create a todo.
      tags:
      - todos
  /todos/:id:
    delete:
      description: delete a single todo by id.
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UpdateOrDeleteTodoRes'
      summary: Delete a single todo.
      tags:
      - todos
    get:
      description: fetch a single todo.
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Get a single todo.
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: update a single todo.
      parameters:
      - description: Todo update data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTodoDTO'
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UpdateOrDeleteTodoRes'
      summary: Update a todo.
      tags:
      - todos
swagger: "2.0"